+++ .gitignore
@@ -1,3 +1,4 @@
 /.parcel-cache
 /dist
 /node_modules
+/raw_data
