+++ package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "clustergrammer-gl": "^0.23.0",
     "igraph-wasm": "^0.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -16,6 +17,7 @@
   },
   "devDependencies": {
     "@parcel/config-default": "^2.3.2",
+    "csv-parser": "^3.0.0",
     "parcel": "^2.3.2",
     "parcel-plugin-static-files-copy": "^2.6.0",
     "parcel-reporter-static-files-copy": "^1.3.4",
